Move collectors to `core` folder
