+++ src/collectors/core/amazon/selectors.ts
