+++ src/collectors/core/carrefour/selectors.ts
