+++ src/collectors/core/intermarche/selectors.ts
