+++ src/collectors/core/laposte/laposte.ts
@@ -1,8 +1,8 @@
-import { ScrapperCollector } from '../scrapperCollector';
+import { ScrapperCollector } from '../../scrapperCollector';
 import { LaposteSelectors } from './selectors';
-import { Driver } from '../../driver/driver';
-import { DownloadedInvoice, Invoice } from '../abstractCollector';
-import { TwofaPromise } from '../../collect/twofaPromise';
+import { Driver } from '../../../driver/driver';
+import { DownloadedInvoice, Invoice } from '../../abstractCollector';
+import { TwofaPromise } from '../../../collect/twofaPromise';
 
 export class LaposteCollector extends ScrapperCollector {
