+++ src/collectors/core/leclerc/selectors.ts
