+++ src/collectors/core/leroy_merlin/selectors.ts
