+++ src/collectors/core/total_energies/selectors.ts
